MSD: Extract shared EnvironmentSwitcherDropdown component
Moves EnvironmentSwitcherDropdown and StagingSiteActionButton into
a shared file, removing the remaining duplication between v1 and v2.
